Style changes and page refactor
