+++ src/components/queue/Broadcast.tsx
@@ -1,20 +1,7 @@
 import { useState } from 'react';
 import { useChannel } from '@ably-labs/react-hooks';
-import {
-  Flex,
-  Input,
-  Button,
-  ModalFooter,
-  Modal,
-  ModalContent,
-  ModalHeader,
-  ModalBody,
-  ModalCloseButton,
-  useToast,
-  Center,
-  useColorModeValue,
-} from '@chakra-ui/react';
-import { DARK_GRAY_COLOR } from '../../utils/constants';
+import { Flex, Input, Button, useToast, useColorModeValue } from '@chakra-ui/react';
+import BroadcastConfirmationModal from '../modals/BroadcastConfirmationModal';
 
 /**
  * Broadcast component that allows staff to broadcast messages everyone
@@ -61,26 +48,12 @@ const Broadcast = () => {
         Broadcast
       </Button>
 
-      <Modal isOpen={isModalOpen} onClose={() => setIsModalOpen(false)}>
-        <ModalContent backgroundColor={useColorModeValue('', DARK_GRAY_COLOR)}>
-          <ModalHeader>Confirm Broadcast</ModalHeader>
-          <ModalCloseButton />
-          <ModalBody>
-            Are you sure you want to broadcast this message? <br /> <br />{' '}
-            <Center>
-              <strong>{broadcastMsg}</strong>
-            </Center>
-          </ModalBody>
-          <ModalFooter>
-            <Button variant='ghost' mr={3} onClick={() => setIsModalOpen(false)}>
-              Cancel
-            </Button>
-            <Button colorScheme='blue' onClick={broadcast}>
-              Confirm
-            </Button>
-          </ModalFooter>
-        </ModalContent>
-      </Modal>
+      <BroadcastConfirmationModal
+        isModalOpen={isModalOpen}
+        setIsModalOpen={setIsModalOpen}
+        broadcastMsg={broadcastMsg}
+        broadcast={broadcast}
+      />
     </Flex>
   );
 };
