+++ src/components/queue/OpenOrCloseQueue.tsx
@@ -1,18 +1,9 @@
 import { useChannel } from '@ably-labs/react-hooks';
-import {
-  Button,
-  ModalFooter,
-  Modal,
-  ModalContent,
-  ModalHeader,
-  ModalBody,
-  ModalCloseButton,
-  useColorModeValue,
-} from '@chakra-ui/react';
+import { Button } from '@chakra-ui/react';
 import { PersonalQueue } from '@prisma/client';
 import { useState } from 'react';
-import { DARK_GRAY_COLOR, DARK_HOVER_COLOR } from '../../utils/constants';
 import { trpc } from '../../utils/trpc';
+import OpenOrCloseQueueModal from '../modals/OpenOrCloseQueueModal';
 
 interface OpenOrCloseQueueProps {
   isQueueOpen: boolean;
@@ -48,7 +39,7 @@ const OpenOrCloseQueue = (props: OpenOrCloseQueueProps) => {
       });
     }
   };
-  
+
   return (
     <>
       <Button
@@ -60,30 +51,12 @@ const OpenOrCloseQueue = (props: OpenOrCloseQueueProps) => {
       >
         {isQueueOpen ? 'Close' : 'Open'} Queue
       </Button>
-
-      <Modal isOpen={isModalOpen} onClose={() => setIsModalOpen(false)}>
-        <ModalContent
-          backgroundColor={useColorModeValue('', DARK_GRAY_COLOR)}
-          boxShadow={`0 0 1px 2px ${DARK_HOVER_COLOR}`}
-        >
-          <ModalHeader>{isQueueOpen ? 'Close' : 'Open'} Queue</ModalHeader>
-          <ModalCloseButton />
-          <ModalBody>Are you sure you want to {isQueueOpen ? 'close' : 'open'} the queue?</ModalBody>
-          <ModalFooter>
-            <Button variant='ghost' mr={3} onClick={() => setIsModalOpen(false)}>
-              Cancel
-            </Button>
-            <Button colorScheme='blue' mr={3} onClick={() => handleOpenOrCloseQueue(false)}>
-              Confirm
-            </Button>
-            {isQueueOpen && (
-              <Button colorScheme='green' onClick={() => handleOpenOrCloseQueue(true)}>
-                Confirm and Clear Queue
-              </Button>
-            )}
-          </ModalFooter>
-        </ModalContent>
-      </Modal>
+      <OpenOrCloseQueueModal
+        isModalOpen={isModalOpen}
+        setIsModalOpen={setIsModalOpen}
+        isQueueOpen={isQueueOpen}
+        handleOpenOrCloseQueue={handleOpenOrCloseQueue}
+      />
     </>
   );
 };
