+++ src/pages/activity/index.tsx
@@ -1,15 +1,17 @@
 import { NextPage } from 'next';
+import dynamic from 'next/dynamic';
 import Layout from '../../components/layout/Layout';
-import { useSession } from 'next-auth/react';
-import ActivityView from '../../components/activity/ActivityView';
 
 /**
- * Activity page which displays the activity log 
+ * Activity page which displays the activity log
  */
 const ActivityPage: NextPage = () => {
-  const { data: session } = useSession();
-
-  return <Layout>{session?.user && <ActivityView user={session.user} />}</Layout>;
+  const ActivityView = dynamic(() => import('../../components/activity/ActivityView'));
+  return (
+    <Layout>
+      <ActivityView />
+    </Layout>
+  );
 };
 
 export default ActivityPage;
