+++ src/pages/create-queue/index.tsx
@@ -1,9 +1,11 @@
 import { NextPage } from 'next';
 import Layout from '../../components/layout/Layout';
 import { UserRole } from '@prisma/client';
-import CreatePersonalQueue from '../../components/queue/CreatePersonalQueue';
+import dynamic from 'next/dynamic';
 
 const CreateQueuePage: NextPage = () => {
+  const CreatePersonalQueue = dynamic(() => import('../../components/queue/CreatePersonalQueue'));
+
   return (
     <Layout restrictedTo={[UserRole.STAFF]}>
       <CreatePersonalQueue />
